backend/pyproject.toml
[project]
name = "shopper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "modal>=1.2.1",
    "Django>=5.1",
    "browser-use",
    "django-cors-headers>=4.6.0",
    "django-q2",
    "psycopg[binary]>=3.2",
    "python-dotenv>=1.0.1",
    "pytest-playwright>=0.7.1",
    "ipykernel>=7.1.0",
]
